services/async_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
orjson>=3.11.1,<3.12

elasticsearch[async]==8.13.2
redis==5.0.4
python-dotenv==1.0.1
pydantic-settings==2.5.2

PyJWT==2.9.0
